mocks: move mocks from mocks_test.go to mocks package
