+++ worker/pricetables_test.go
@@ -7,30 +7,30 @@ import (
 	"time"
 
 	"go.sia.tech/renterd/api"
+	"go.sia.tech/renterd/internal/test/mocks"
 )
 
 func TestPriceTables(t *testing.T) {
 	// create host & contract stores
-	hs := newHostStoreMock()
-	cs := newContractStoreMock()
+	hs := mocks.NewHostStore()
+	cs := mocks.NewContractStore()
 
 	// create host manager & price table
 	hm := newTestHostManager(t)
 	pts := newPriceTables(hm, hs)
 
 	// create host & contract mock
-	h := hs.addHost()
-	c := cs.addContract(h.hk)
+	h := hs.AddHost()
+	c := cs.AddContract(h.PublicKey())
 
-	cm := &chainMock{
-		cs: api.ConsensusState{
-			BlockHeight: 1,
-		},
-	}
+	cm := mocks.NewChain(api.ConsensusState{
+		BlockHeight: 1,
+	})
+	cState, _ := cm.ConsensusState(context.Background())
 
 	blockHeightLeeway := 10
 	gCtx := WithGougingChecker(context.Background(), cm, api.GougingParams{
-		ConsensusState: cm.cs,
+		ConsensusState: cState,
 		GougingSettings: api.GougingSettings{
 			HostBlockHeightLeeway: blockHeightLeeway,
 		},
@@ -39,7 +39,7 @@ func TestPriceTables(t *testing.T) {
 	// expire its price table
 	expiredPT := newTestHostPriceTable()
 	expiredPT.Expiry = time.Now()
-	h.hi.PriceTable = expiredPT
+	h.UpdatePriceTable(expiredPT)
 
 	// manage the host, make sure fetching the price table blocks
 	fetchPTBlockChan := make(chan struct{})
@@ -50,7 +50,7 @@ func TestPriceTables(t *testing.T) {
 	}))
 
 	// trigger a fetch to make it block
-	go pts.fetch(gCtx, h.hk, nil)
+	go pts.fetch(gCtx, h.PublicKey(), nil)
 	time.Sleep(50 * time.Millisecond)
 
 	// fetch it again but with a canceled context to avoid blocking
@@ -58,14 +58,14 @@ func TestPriceTables(t *testing.T) {
 	// update
 	ctx, cancel := context.WithCancel(gCtx)
 	cancel()
-	_, _, err := pts.fetch(ctx, h.hk, nil)
+	_, _, err := pts.fetch(ctx, h.PublicKey(), nil)
 	if !errors.Is(err, errPriceTableUpdateTimedOut) {
 		t.Fatal("expected errPriceTableUpdateTimedOut, got", err)
 	}
 
 	// unblock and assert we paid for the price table
 	close(fetchPTBlockChan)
-	update, _, err := pts.fetch(gCtx, h.hk, nil)
+	update, _, err := pts.fetch(gCtx, h.PublicKey(), nil)
 	if err != nil {
 		t.Fatal(err)
 	} else if update.UID != validPT.UID {
@@ -74,8 +74,8 @@ func TestPriceTables(t *testing.T) {
 
 	// refresh the price table on the host, update again, assert we receive the
 	// same price table as it hasn't expired yet
-	h.hi.PriceTable = newTestHostPriceTable()
-	update, _, err = pts.fetch(gCtx, h.hk, nil)
+	h.UpdatePriceTable(newTestHostPriceTable())
+	update, _, err = pts.fetch(gCtx, h.PublicKey(), nil)
 	if err != nil {
 		t.Fatal(err)
 	} else if update.UID != validPT.UID {
@@ -83,21 +83,21 @@ func TestPriceTables(t *testing.T) {
 	}
 
 	// manually expire the price table
-	pts.priceTables[h.hk].hpt.Expiry = time.Now()
+	pts.priceTables[h.PublicKey()].hpt.Expiry = time.Now()
 
 	// fetch it again and assert we updated the price table
-	update, _, err = pts.fetch(gCtx, h.hk, nil)
+	update, _, err = pts.fetch(gCtx, h.PublicKey(), nil)
 	if err != nil {
 		t.Fatal(err)
-	} else if update.UID != h.hi.PriceTable.UID {
+	} else if update.UID != h.PriceTable().UID {
 		t.Fatal("price table mismatch")
 	}
 
 	// refresh the price table on the host and make sure fetching doesn't update
 	// the price table since it's not expired
-	validPT = h.hi.PriceTable
-	h.hi.PriceTable = newTestHostPriceTable()
-	update, _, err = pts.fetch(gCtx, h.hk, nil)
+	validPT = h.PriceTable()
+	h.UpdatePriceTable(newTestHostPriceTable())
+	update, _, err = pts.fetch(gCtx, h.PublicKey(), nil)
 	if err != nil {
 		t.Fatal(err)
 	} else if update.UID != validPT.UID {
@@ -107,13 +107,13 @@ func TestPriceTables(t *testing.T) {
 	// increase the current block height to be exactly
 	// 'priceTableBlockHeightLeeway' blocks before the leeway of the gouging
 	// settings
-	cm.cs.BlockHeight = validPT.HostBlockHeight + uint64(blockHeightLeeway) - priceTableBlockHeightLeeway
+	cm.UpdateHeight(validPT.HostBlockHeight + uint64(blockHeightLeeway) - priceTableBlockHeightLeeway)
 
 	// fetch it again and assert we updated the price table
-	update, _, err = pts.fetch(gCtx, h.hk, nil)
+	update, _, err = pts.fetch(gCtx, h.PublicKey(), nil)
 	if err != nil {
 		t.Fatal(err)
-	} else if update.UID != h.hi.PriceTable.UID {
+	} else if update.UID != h.PriceTable().UID {
 		t.Fatal("price table mismatch")
 	}
 }
